pyproject.toml
[tool.poetry]
name = "felipebastosprobr"
version = "0.1.0"
description = "Meu site pessoal"
authors = ["Felipe Bastos Nunes <felipebastos@ifce.edu.br>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
djangorestframework-simplejwt = "^5.1.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pylint = "^2.13.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
